Begin to split up Requirements component.
